Reworked according to review
